Split PostgisGeographyHandler into it's own file, and factor our shared read code into a base PostgisHandler class
